chore: moving the query builder to the elasticsearch module
